requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Graph database
neo4j>=5.18.0

# Cache & queue
redis>=5.0.0
celery>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.1.0
python-multipart>=0.0.9

# HTTP client (for connectors)
httpx>=0.27.0

# SQL parsing (for connector lineage extraction)
sqlglot>=23.0.0

# JSON Schema validation
jsonschema>=4.21.0

# PostgreSQL connector
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
